Fin TP Integrador

+++ src/main/java/com/ar/bankingonline/api/dtos/TransferDto.java
@@ -10,9 +10,9 @@ public class TransferDto {
 
     private Long id;
 
-    private int origin;
+    private Long origin;
 
-    private int target;
+    private Long target;
 
     private Date date;
